Allow overriding workflow mode with squash using a flag
When the mode in the config is 'ITERATIVE', sometimes users
want to run the workflow once in 'SQUASH' to fix an issue
or to run an initial migration. In other cases users are running
a workflow that is defined in some Starkark library that
they don't own, but is shared by many teams.

For those cases, it's more convenient to run their workflow
as:

copybara path/to/copy.bara.sky workflow_name some_ref --squash

instead of having to edit the config locally and run it.

BUG=123405656
PiperOrigin-RevId: 231234010
Change-Id: I443ecc7bb70656ebe18dcd8eeb6d1d7d8d4dabd0