Allow notify options in Gerrit destination
This is an option that gets appended to the reference:

refs/for/master%notify=NONE

As it has implications in Destination.describe(), that
is used to introspect workflows, instead of hard-coding
it in the ref, I'm taking advantage of
WriteHook#getPushReference() which already has logic
for generating the options for reviewers.

Additionally, I'm moving the logic for the topic option
also to this method, as it's consistent with the rest
and won't affect describe() either.

BUG=123536947
PiperOrigin-RevId: 231265395
Change-Id: I137af4523ce1c33640b2914422055e1521176e17