Fix an error-prone issue with VE varargs
When no args are passed to checkCondition, it is very common that the
error contains arbitrary characters (like %s). This is problematic as
we try to interpret it as format arguments.

This change disables String.format for zero arguments. The risk of
this change is that we could potentially see some '%s' in errors
instead of the argument if a user forgets to add the argument. But
the risk is much smaller and the consequences too.

FIXED=123689132

Change-Id: Id932d08c2edf43d8211d1dd8842c1512b4568943