Clean untracked directories in git.destination
When we have untracked directories (instead of files), we weren't
properly cleaning the workdir. As we were only doing 'git clean -f',
without passing -d (For untracked directories).

In general, this doesn't create much harm, as we don't care about
files left in the workdir. But in the case of git.gerrit_destination
when used with allow_empty_diff_patchset = False (check that we
don't create an empty patchset), we do additional checkouts and
commit work that needs that the workdir is clean.

This fix adds a function in GitRepository, since we had the same
issue in git integrate code.

FIXED=123830784

Change-Id: Icb04e83abaa306c19c95b1fef084f5ac72cb36cf