Add support for after all migrations hook
Currently we have hooks executed per change migrated. That means that
for ITERATIVE mode we execute them N times. For some cleanup tasks
that we want to run, we'd like to run something just at the end
instead.

As part of this change, I extracted the event notification code to the
workflow so that we can record the events locally too. The code is a
bit more complex (uses a Consumer), but this guarantees that the
life-cycle of the accumulator is the correct one (Even if the workflow
is reused in one invocation, something we could do in the future).

PiperOrigin-RevId: 236129260
Change-Id: Id315e105230b188b6c9665c07bc29da6bd905739