Add delete_reference to GitHub API and Starlark
We also ensure that 'refs/heads/master' is not deleted. Since there is
no good reason for doing that with Copybara.

In addition I made update_reference to work with complete refs instead
of short version. This is inconsistent with get_reference and its own
documentation wasn't correct. For delete we also want to be able to
delete tags for example (if needed). So it is made sense to move
everything to complete ref.

Short ref is still accepted (with warning) for update_reference for
compatibility reasons. But will be replaced by a check soon.

PiperOrigin-RevId: 237333837
Change-Id: I5e68791023a23d37019479af190b93cd30a79616