Be able to query for Pull Requests in Starlark
This exposes Pull Requests in Starlark with the options for sort,
direction, etc.

As part of this change state for the PR is returned as upper case,
since we usually use uppercase for enums and would be weird to
query like get_pull_requests(state = "OPEN") and get back a result
that has state "open".

Change-Id: Ib34640d3aed29f04ef96bab142ec259b344c023e