Delete PR branches in GitHub destination
Currently, when a user sets pr_branch_to_update = True, Copybara can
automatically merge existing presubmit PRs. This PRs are usually
created for running tests. One of the effects is that it leaves
the references in the main repo.

The ideal solution would be to use a fork for the account doing the
presubmit change. But because of some internal limitations, we need
to use the main repo for now.

This change allows to delete those reference (but not the PR) after
merged. The order is important. We need to do it after push so that
GitHub marks the PR as merged before.

In order to use it, a user that already uses pr_branch_to_update,
will set delete_pr_branch = True field or use the
--github-destination-delete-pr-branch flag.

BUG=122672107
Change-Id: I6ccc0660054484ff08d6ea4f5b4fa3095bb1e1a9