Fix submodules when path not equal to name
The current implementation uses submodule--helper. An internal command
that returns the paths affected by submodules. We were using that path
as the name of the submodule. But that is not true. The name can be
"foo" but fetched in "subfolder/foo".

This implementation gets rid of the submodule--helper (good change in
its own as it is an internal command) and replaces by standard git
config usage. We have to parse a bit the output but similar or better
than the output of submodule--helper.

FIXED=129353500
Change-Id: I06957fa34933f43902c063fb54542bfd215de23e