Decouple Gerrit change resolution from fetching
This is the first change for being able to use the API in
gerrit.origin. We will just resolve the ref to a change
and then use the API to check if the branch matches a
branch config field.

I want to further refactor this code and remove the Gerrit
stuff from GitRepoType. But not in this change.

Change-Id: I15e59db0ad35cc6b9c5bec6e5ab553a4464fb9b5