Detect if rebase nullifies the change
When a presubmit workflow is run with a baseline, if a
HEAD config is used, it might generate differences that
make the commit not empty (Updated transformations that
were done in unrelated changes). If those changes are
migrated to the main branch, they are already applied.
When we rebase, git will see that everything affected
is already in the main branch and remove our commit.

The effect for this is mild for git.destination but ends
up in an error for git.github_pr_destination. in PR
destination we try to create a PR and it will fail because
there is no difference between main branch and the
PR branch.

Sample stack trace of the failure:

        at com.google.copybara.git.github.api.GitHubApi.createPullRequest(GitHubApi.java:192)
        at com.google.copybara.git.GitHubPrDestination$1.write(GitHubPrDestination.java:217)
        at com.google.copybara.WorkflowRunHelper$ChangeMigrator.doMigrate(WorkflowRunHelper.java:625)
        at com.google.copybara.WorkflowRunHelper$ChangeMigrator.migrate(WorkflowRunHelper.java:395)
        at com.google.copybara.WorkflowMode.runChangeRequest(WorkflowMode.java:395)
        at com.google.copybara.WorkflowMode.access$500(WorkflowMode.java:61)
        at com.google.copybara.WorkflowMode$4.run(WorkflowMode.java:297)
        at com.google.copybara.Workflow.run(Workflow.java:263)
        at com.google.copybara.Copybara.run(Copybara.java:87)
        at com.google.copybara.MigrateCmd.run(MigrateCmd.java:53)
        at com.google.copybara.Main.runInternal(Main.java:228)
        at com.google.copybara.Main.run(Main.java:120)

FIXED=129366965
Change-Id: I72f6e6129c20b3a1797cb4f20113250002e2094b