Show a list of all migrations in info
Currently copybara info shows information of each migration status.
If no workflow is pass it uses 'default'. This change removes that
functionality and instead shows the list of migrations when no
workflow is passed.

We use the newly introduced 'description' field to show an
informative text about what the workflow does.

FIXED=129785304
Change-Id: Ia9dd0b5c210d1f0a775235e26b2b4ccb18ba933b