Introduce ShowDiff in origin
Show the difference between lastRev and curRev at origin to let users
see what they are going to import in the early stage of a workflow,
compared to checking the difference at destination.

Add flag --diff-in-origin. When set, only show diff in origin.
Currently, this only works for git source control.

PiperOrigin-RevId: 243302813
Change-Id: I8ac95a7fe84a4f7b2f42f6194a04570eb514d65c