Allow disabling checkout for core.workflow
Currently core.workflow always does a checkout of the origin.
Normally this is what the user wants, as the purpose of core.workflow
is to move files from the origin to destination.

But there is a use case of Copybara where core.workflow is used to
update a dependency in your repo. For example Foo (owned project)
depends on Bar project (external). Foo has a file that contains a
reference to the version of Bar to use (Maven package, Bazel version,
etc.). We want to update the version of our own file using the
dependency version metadata, but we are not interested in the files
(As the build tool itself downloads the files when needed).

This change allows to disable the checkout.

Change-Id: Id87b9aabc1aacd51b81131e8cd605d612284fbfa