Allow empty change exception in Starlark transforms
Currently computing if a change should or shouldn't be migrated or if
a change is empty is computed in Java. That works well for most of the
cases, but sometimes we want to be able to stop a migration for
arbitrary reasons (a label not present, etc.). We can create a
ValidationException error, but the problem with that is that the
exit code will be 2. This interact bad with other automated services
since we cannot differentiate between a real error or a "don't
process this change" error.

This change allows us to exit with exit code 4 (noop).

Change-Id: I7286088a8bc8159c4b9bd8af8c20bb76cf23dd71