Add a force flag for github_pr_origin
Currently there are several filters to not migrate changes:
open/closed, labels, etc. But we don't have a way of forcing
the import if any of these filters do not match. Sometimes
for example we want to be able to import closed PRs.

This change adds a flags to allow imports of PRs that the
user knows should be imported.

FIXED=134928194

Change-Id: I627e85bd83d56bca8427c4aa589c1a59016d2359