Replace $ with \$ when doing appendReplacement
Currently if a $ ends up in todo_replace and filter_replace
replacement, we fail internally because head version of re2j
has added support for groups.

This fixes the issue for these transformations. core_replace
is not affected by this issue, as the pattern that we use
comes from the 'after' field.

FIXED=135633121
PiperOrigin-RevId: 254253851
Change-Id: Iffc134c8f4b89673daff5c47217cecc102c3d93d