Add delegate.ask() for delegate consoles
These consoles should delegate to their inner console instead of
relying on the default (throw exception).

The problem happened in the debugger, as the debugger calls ask()
but if we wrap the console with a delegate, it throw unsupported
exception.

Added the class type for debugging in base class. Not the best
user friendly message but will help to debug if happens again.

PiperOrigin-RevId: 256290515
Change-Id: Ia6efff6b1fa69321397b96a4a493e6535f4c662f