Honor origin_files for smart_prune baseline
Currently, when smart_prune is used, the baseline checkout doesn't
honor excluded files (or not included paths) during the checkout
used for getting a better diff.

This is not a problem normally as smart_prune only cares about files
changed in head and only uses baseline to filter even more.

But there is a case where this doesn't work: If there is a
transform that fails because of the presence of those files. For
example a dynamic transform that acts on a glob or a verify_no_match.

This cl extracts the common logic for checkout and calls it from
both places.

FIXED=140950556

Change-Id: Ibcfb116b73082cac97e2a720ecef89fbff3838aa