remove unnecessary uses of Location
The SkylarkCallable(structField=true) feature was never supposed to support
useLocation=true, but sadly we forgot to enforce this. This change removes
a couple of places where the Location was used to throw an EvalException.

It is legal to throw EvalException from a Starlark built-in function;
the interpreter will add the location to the exception automatically.
The Starlark.errorf utility function makes this very convenient.

PiperOrigin-RevId: 286618143
Change-Id: Iaa6b548eb3c7bc0329dbfd47278c63057c147dab