Allow interactive merge for rebase
Supports interactive rebase error message and lock dir

We currently assume git-am as rebase algorithm, but this is not likely to be the
case going forward; the git debian experimental branch switched the default.

The --am flag to force git-am is not understood by current git versions,
so setting it does not work either.

For now, this just changes the error message parsing and cleans up the the
interactive rebase dir. Eventually, we probably want to force one algorithm to
yield consistent results.

PiperOrigin-RevId: 288345499
Change-Id: I6f15c0c97c6e49f4d33803666dcf1fb7969197e3