Fetch SHA-1 refs from remotes without HEAD
Currently if we try to fetch a SHA-1 from a remote without HEAD, we
first try to fetch the default ref and look the the commit there. The
problem is that some remotes don't define HEAD, making this fetch to
fail and throw an Exception. We never try fetching the SHA-1 directly.

This change makes the HEAD fetch tolerant to failure and still tries
to fetch by SHA-1 on failure.

FIXED=156040827
Change-Id: Ib7e5c03ef435bad4e9d101b08910bd2e4158e320