Expose current revision date & time in Starlark
Currently we expose many labels about the current change. Like the
current message or title. But we don't expose the date. We do for
each change in ctx.changes.current. But sometimes we don't have a
history (--force is being used).

This exposes COPYBARA_CURRENT_REV_DATE_TIME label. It is in ISO format.
For example: "2020-05-28T21:25:00Z". Then users can chop the part they
need.

Change-Id: Ifdda5a2ba3e173273379be2646be2a21c5c84208