add --to-folder testing flag
Many users want to test the outcome of a migration to test that
what Copybara generates is correct. This results on users
temporarily modifying their workflows or worse, add a permanent
testing workflow.

git.destination has a mode that allows to put the change in
a local repository. But sometimes is hard to work with it and
has some non-trivial bugs.

This change add a Workflow flag, --to-folder, that simply replaces
the destination with a folder.destination for any core.workflow.

Note that any assumption that the destination is the configured
destination might fail. For example post-migration hooks.

This flag is best used with --squash.

Change-Id: Iae330a584e85ffdf105ab0258f291cded7a1b13c