gerrit.dest: Use hashtag for reusing changes [1/N]
Copybara tries to reuse changes for presubmit and post-submit (in
squash mode). For gerrit.destination, the algorithm is to generate
Change-Ids based on the workflow identity until we find none or
a NEW change (NEW = Active review). This does up to 100 queries
for different change-Ids and fails if it cannot find an available
one.

Note that this happens on very particular scenarios because of
the way the workflow identity works. In practice this is a problem
only in postsubmit that uses SQUASH and gerrit.destination (With
submit = False). This is not common. But there are cases that
it is desirable (For example our own Bazel dependency update
workflow and some other internal usages).

This change is the first of many that will use hashtags instead.
The idea is that we will always push to gerrit with an internal
hashtag. Then in the following executions of Copybara we will
query if there is any open change that we can reuse.

This first change only tries to reuse the Change-Id using hashtags
for query. But if not found it still uses the existing Change-Id
search algorithm. But on push it will add the hashtag. In this
way, after some time (1 month), we can change the algorithm to
generate random Change-id without affecting our users (Since
most of the changes after a month will have hashtag).

BUG=159696870

Change-Id: I41ae929ccaecf1131e8c1739fe95f9ef1db228a6