gerrit: Use max change number for reuse
Currently when we find more than one change number associated
with one hashtag, we use the first element in the list returned
by Gerrit. The problem with this approach is that Gerrit doesn't
return its results sorted. So we end up reusing one of the other.

This is not a big deal, as it is very rare to create two changes
with the same hashtag, and it is debatable which one is best. We
chose the max number consistently now.

Change-Id: I50369bbf7222171eea629d3fe6274e189debddf1